Remove deprecated `threaded`-parameter
The `threaded`-parameter for SleekXMPP got deprecated several years ago,
probably because it got replaced by the `block`-parameter. In any case
its name was misleading, as all it does is to control whether the event
dispatcher will run in a separate thread or not and not if threads are
used at all or not (hint: they are).