Let EcheLOn reply properly to IQs
Previously EcheLOn created completely new IQs for each response to an
IQ. That caused new ids to be used, which shouldn't be the case. This
commit fixes that.

The problem still exists in XpartaMuPP, but it's harder to fix there,
because it either broadcasts messages or relays them trough EcheLOn.