Return correct stanza id for requests to EcheLOn
For all requests made from clients to XpartaMuPP, which get forwarded to
EcheLOn, return the correct stanza id to the client.

This is implemented using a size limited dictionary for the mapping of
ids between clients and XpartaMuPP and ids between XpartaMuPP and
EcheLOn so ids without a response from EcheLOn get evicted over time
while avoiding an ever growing memory consumption of XpartaMuPP in such
a case.