Change handling for offline EcheLOn
Previously EcheLOn was assumed offline if he wasn't present in the MUC
nick list maintained by XpartaMuPP. This was prone to race conditions on
EcheLOn joining or leaving the room. Now instead of assuming if EcheLOn
is online or not a IQ query is sent in every case and in case EcheLOn
isn't online the error response is handled appropriately.

Additionally this commit sends responses back to the 0ad clients in case
EcheLOn is offline to better comply with the request/response model of
IQ queries. These responses are currently empty result responses, as 0ad
shows nasty error messages in the chat window for error responses.