Fix memory leak for incomplete game reports
A game is only rated if a report from every participant is received.
Until all such reports are received, they're cached in the
`ReportManager.interim_report_tracker` list attribute. Similary JIDs
of the participating players are cached in
`ReportManager.interim_jid_tracker`. As there is no mechanism to
remove items from these lists, if not all participants of a game
submitted a report, it's possible to use that to DoS EcheLOn.

Such a DoS simply involves sending lots of game reports for rated
games where only one of the two players submits a game report. That'll
lead EcheLOn to use more and more memory until there is no more
memory available.

This commit replaces the use of a lists for `interim_report_tracker`
and `interim_jid_tracker` with `collections.deque` with a `maxlen`
of 4096 elements, causing old unfinished games to be dropped, if the
deque is full.