Untangle XpartaMuPP and EcheLOn
Up until now XpartaMuPP was able to relay requests meant for EcheLOn to
EcheLOn. As 0ad now sends such requests directly to EcheLOn we don't
need that functionality anymore, which makes the code way more readable.

Additionally this closes a vulnerability which allowed clients to
manipulate the ratings of other players by submitting requests directly
to EcheLOn, but including the <sender> in the sent XML, which normally
only XpartaMuPP did for requests forwarded to EcheLOn.