Let the bots answer with meaningful messages
As we have two separate bots with clearly defined responsibilities now,
it adds a nice touch to let them tell what their purpose is, when
they're asked.

As part of that "Ratings" got a nick change to "RatingsBot" to not
trigger the message accidentally when talking about ratings and "WFGbot"
got a nick change to "WFGBot" to be in line with "RatingsBot" in terms
of capitalization.