Limit the max amount of open games to 128
XpartaMuPP was vulnerable to DoS, by opening an infinitive amount of
games, which causes the responses to occupy all available memory and
make the lobby unusable, because the responses become too big and to
compute intensive.

Even 128 might be still too high for a useable lobby, but at least it's
not possible anymore for clients to use it to fill up all available
memory. The proper fix for all related performance problems is PubSub,
which will be introduced later.