Use MUC-plugin provided participant information
It turns out the XEP-0045 plugin for SleekXMPP already stores
information about all participants in a MUC room. So instead of
collecting and storing this information in the bots again, simply rely
on the information provided by the plugin. This reduces the required
memory for the additional storage of nick-JID-relations in the bots and
reduces the code complexity.